добавлена базовая обработка платежа, обработка ошибок, хранение истории транзакций

+++ database/migrations/2020_01_18_071942_create_transactions_table.php
@@ -16,10 +16,10 @@ class CreateTransactionsTable extends Migration
         Schema::create('transactions', function (Blueprint $table) {
             $table->bigIncrements('id');
 
-            $table->unsignedInteger('sender_id')->nullable();
+            $table->unsignedBigInteger('sender_id')->nullable();
             $table->string('sender_name')->nullable();
 
-            $table->unsignedInteger('receiver_id')->nullable();
+            $table->unsignedBigInteger('receiver_id')->nullable();
             $table->string('receiver_name')->nullable();
 
             $table->double('amount');
